src/requirements.txt
PyNaCl
python-dotenv
pycparser
discord.py